test(signing): keep the shared fixtures in their own file
newTestKey is used by both test files but sat above the keymanager-only HTTP
fixtures, where it reads as keymanager's. Moving every fixture into
fixtures_test.go leaves each test file matching one source file, and answers
where the next fixture goes without another round of this question.

Only declarations moved. encoding/base64 and encoding/json leave
keymanager_test.go with them.
